requirements.txt
tkinter
scikit-learn
numpy
matplotlib
Pillow